feat(api)(frontend): add task scheduling feature (#27)
This PR introduces the concept of tasks within the API module. A use case for task is running an ingestion pipeline from a datasource.

The result of this PR is that we will not need to current uniform scheduling mechanism. Rather, users are able to pick when to schedule the ingestion of a specific datasource.

Closes of #8